Remove a call to Type.GetProperty to improve AOT support (#2048)
EPPlus is mostly usable in a native AOT compilation context. One exception to
this support is the use of the reflection method Type.GetProperty in
TypeCompat.GetPropertyValue.

Fortunately, there is only one place where this method is used, in
ExcelVBACollectionBase. By adding a type constraint on the generic type T of
ExcelVBACollectionBase, we can get rid of the use of reflection.

In practice, an AOT compiled project that uses EPPlus to read an Excel file
that contains VBA macro could throw NullReferenceExceptions.